Code update and file rename
Fix: clean up and improving code using all the comments.
Fix: rename file to a more meaning and more easy to search
